Extract CourseVacancyStatus and CourseStudyModes classes
