+++ spec/services/find_sync/sync_provider_from_find_spec.rb
@@ -351,137 +351,4 @@ RSpec.describe FindSync::SyncProviderFromFind, sidekiq: true do
       end
     end
   end
-
-  describe 'CourseVacancyStatus#derive' do
-    context 'when study_mode is part_time' do
-      let(:study_mode) { 'part_time' }
-
-      [
-        { description: 'no_vacancies', vacancy_status: :no_vacancies },
-        { description: 'both_full_time_and_part_time_vacancies', vacancy_status: :vacancies },
-        { description: 'full_time_vacancies', vacancy_status: :no_vacancies },
-        { description: 'part_time_vacancies', vacancy_status: :vacancies },
-      ].each do |pair|
-        it "returns #{pair[:vacancy_status]} when description is #{pair[:description]}" do
-          derived_status = SyncCoursesFromFind::CourseVacancyStatus.new(
-            pair[:description],
-            study_mode,
-          ).derive
-
-          expect(derived_status).to eq pair[:vacancy_status]
-        end
-      end
-
-      it 'raises an error when description is an unexpected value' do
-        expect {
-          SyncCoursesFromFind::CourseVacancyStatus.new('foo', study_mode).derive
-        }.to raise_error(
-          SyncCoursesFromFind::CourseVacancyStatus::InvalidFindStatusDescriptionError,
-        )
-      end
-    end
-
-    context 'when study_mode is full_time' do
-      let(:study_mode) { 'full_time' }
-
-      [
-        { description: 'no_vacancies', vacancy_status: :no_vacancies },
-        { description: 'both_full_time_and_part_time_vacancies', vacancy_status: :vacancies },
-        { description: 'full_time_vacancies', vacancy_status: :vacancies },
-        { description: 'part_time_vacancies', vacancy_status: :no_vacancies },
-      ].each do |pair|
-        it "returns #{pair[:vacancy_status]} when description is #{pair[:description]}" do
-          derived_status = SyncCoursesFromFind::CourseVacancyStatus.new(
-            pair[:description],
-            study_mode,
-          ).derive
-
-          expect(derived_status).to eq pair[:vacancy_status]
-        end
-      end
-
-      it 'raises an error when description is an unexpected value' do
-        expect {
-          SyncCoursesFromFind::CourseVacancyStatus.new('foo', study_mode).derive
-        }.to raise_error(
-          SyncCoursesFromFind::CourseVacancyStatus::InvalidFindStatusDescriptionError,
-        )
-      end
-    end
-  end
-
-  describe 'CourseStudyModes#derive' do
-    context 'when the course has no course options' do
-      let(:course) { create(:course) }
-
-      it 'returns both study modes if the course supports both study modes' do
-        course.full_time_or_part_time!
-
-        study_modes = SyncCoursesFromFind::CourseStudyModes.new(course).derive
-        expect(study_modes).to match_array %w[full_time part_time]
-      end
-
-      it 'returns one study mode if the course only supports one' do
-        course.full_time!
-
-        study_modes = SyncCoursesFromFind::CourseStudyModes.new(course).derive
-        expect(study_modes).to match_array %w[full_time]
-      end
-    end
-
-    context 'when the course has existing course options with uniform study modes' do
-      let(:course) do
-        create(:course, :part_time) do |course|
-          course.course_options = [
-            create(:course_option, :part_time, course: course),
-            create(:course_option, :part_time, course: course),
-            create(:course_option, :part_time, course: course),
-          ]
-        end
-      end
-
-      it 'returns the existing study mode' do
-        study_modes = SyncCoursesFromFind::CourseStudyModes.new(course).derive
-        expect(study_modes).to match_array %w[part_time]
-      end
-
-      it 'returns both study modes if the course changes to support both study modes' do
-        course.full_time_or_part_time!
-
-        study_modes = SyncCoursesFromFind::CourseStudyModes.new(course).derive
-        expect(study_modes).to match_array %w[full_time part_time]
-      end
-
-      it 'returns both study modes if the course changes from one to the other' do
-        course.full_time!
-
-        study_modes = SyncCoursesFromFind::CourseStudyModes.new(course).derive
-        expect(study_modes).to match_array %w[full_time part_time]
-      end
-    end
-
-    context 'when the course has existing course options with a mix of study modes' do
-      let(:course) do
-        create(:course, :with_both_study_modes) do |course|
-          course.course_options = [
-            create(:course_option, :part_time, course: course),
-            create(:course_option, :part_time, course: course),
-            create(:course_option, :full_time, course: course),
-          ]
-        end
-      end
-
-      it 'returns both study modes' do
-        study_modes = SyncCoursesFromFind::CourseStudyModes.new(course).derive
-        expect(study_modes).to match_array %w[full_time part_time]
-      end
-
-      it 'returns both study modes even if the course changes to a specific one' do
-        course.full_time!
-
-        study_modes = SyncCoursesFromFind::CourseStudyModes.new(course).derive
-        expect(study_modes).to match_array %w[full_time part_time]
-      end
-    end
-  end
 end
